Declare an explicit type for colours.
An imperial buttload of places assumes these to be 8 bits, which makes it
impossible to add support for 256 (or more!) colours.  This commit tries
to allow expanding it in the future; sadly, it is badly incomplete.  At
least, new code can use the new type.

Also, beware of "branded" colours which use bits above 8 for inverse,
underline, etc.

+++ crawl-ref/source/dbg-scan.cc
@@ -30,7 +30,7 @@ static void _dump_item(const char *name, int num, const item_def &item)
          item.plus, item.plus2, item.special);
 
     mprf("    quant: %d; colour: %d; ident: 0x%08"PRIx32"; ident_type: %d",
-         item.quantity, item.colour, item.flags,
+         item.quantity, (int)item.colour, item.flags,
          get_ident_type(item));
 
     mprf("    x: %d; y: %d; link: %d", item.pos.x, item.pos.y, item.link);
